Moved Window to utilities module
